Home page UI added
